+++ src/imio/project/pst/subscribers.py
@@ -6,10 +6,10 @@ from eea.facetednavigation.criteria.interfaces import ICriteria
 from imio.helpers.content import set_to_annotation
 from imio.pm.wsclient.browser.settings import notify_configuration_changed
 from imio.project.core.events import empty_fields
-from imio.project.pst.browser.controlpanel import field_constraints
+from imio.project.pst.content.pstprojectspace import field_constraints
 from imio.project.pst.content.action import IPSTSubAction
 from imio.project.pst.interfaces import IActionDashboardBatchActions
-from imio.project.pst.interfaces import IImioPSTProject
+from imio.project.core.content.projectspace import IProjectSpace
 from imio.project.pst.interfaces import IOODashboardBatchActions
 from imio.project.pst.interfaces import IOSDashboardBatchActions
 from imio.project.pst.interfaces import ITaskDashboardBatchActions
@@ -52,15 +52,8 @@ def wsclient_configuration_changed(event):
 
 
 def projectspace_created(obj, event):
-    """When a projectspace is created, if it is not the PST, we constrain types to ('project', )"""
-
+    """  """
     alsoProvides(obj, IOSDashboardBatchActions)
-
-    if not IImioPSTProject.providedBy(obj):
-        behaviour = ISelectableConstrainTypes(obj)
-        behaviour.setConstrainTypesMode(1)
-        behaviour.setLocallyAllowedTypes(['project', ])
-        behaviour.setImmediatelyAddableTypes(['project', ])
 
 
 def projectspace_moved(obj, event):
@@ -91,6 +84,26 @@ def projectspace_moved(obj, event):
             logger.info('Replaced default col {} by {} on {}'.format(old_uid, default_col, ob.absolute_url()))
         else:
             raise ValueError("Cannot update default col on {}".format(ob.absolute_url()))
+
+
+def pstprojectspace_modified(obj, event):
+    """
+      Handler when pstprojectspace modified
+      Update :
+      - customViewFields defined on DashboardCollection when projects columns modified
+      - pstsubaction_fields when pstaction_fields modified
+    """
+    if not event.descriptions:
+        return
+    for desc in event.descriptions:
+        for attr in desc.attributes:
+            if attr.endswith('columns'):
+                pc = api.portal.get_tool('portal_catalog')
+                ps_path = '/'.join(obj.getPhysicalPath()) + '/{}'.format(attr[:-8])
+                for brain in pc(path={'query': ps_path, 'depth': 1}):
+                    brain.getObject().customViewFields = tuple(getattr(obj, attr))
+            if attr == 'pstaction_fields':
+                obj.pstsubaction_fields = obj.pstaction_fields
 
 
 def strategic_created(obj, event):
@@ -179,55 +192,5 @@ def registry_changed(event):
     """  """
     if IRecordModifiedEvent.providedBy(event):
         if event.record.interfaceName == 'imio.project.pst.browser.controlpanel.IImioPSTSettings':
-            # we copy new pstaction_fields value in pstsubaction_fields
-            if event.record.fieldName == 'pstaction_fields':
-                api.portal.set_registry_record('imio.project.settings.pstsubaction_fields', event.newValue)
             if event.record.fieldName == 'pstaction_budget_states':
                 api.portal.set_registry_record('imio.project.settings.pstsubaction_budget_states', event.newValue)
-            # we check if we have to remove/add a column from/to a dashboard
-            if event.record.fieldName in ('strategicobjective_fields', 'operationalobjective_fields',
-                                          'pstaction_fields'):
-                if event.oldValue is None:
-                    return  # site creation: nothing to do
-                pt = event.record.fieldName[:-7]
-                ovs, nvs = set(event.oldValue), set(event.newValue)
-                removed = ovs - nvs
-                added = nvs - ovs
-
-                def find_collections(intf):
-                    ret = []
-                    for fld_b in api.content.find(object_provides=intf.__identifier__, portal_type='Folder'):
-                        for col_b in api.content.find(**{'path': {'query': fld_b.getPath()},
-                                                         'portal_type': 'DashboardCollection'}):
-                            ret.append(col_b.getObject())
-                    return ret
-
-                def add_rm_col(aset, action):
-                    for fld in aset:
-                        if fld not in FIELDS_COLUMNS[pt]['fields']:
-                            continue
-                        col_name = FIELDS_COLUMNS[pt]['fields'][fld]
-                        for collection in find_collections(FIELDS_COLUMNS[pt]['intf']):
-                            nl = list(collection.customViewFields)
-                            if action == 'rm' and col_name in nl:
-                                nl.remove(col_name)
-                                collection.customViewFields = tuple(nl)
-                            elif action == 'add' and col_name not in nl:
-                                i = 0
-                                for column in COLUMNS_FOR_CONTENT_TYPES[pt]:
-                                    if column != col_name:
-                                        try:
-                                            i = nl.index(column)  # find the previous column...
-                                        except ValueError:
-                                            continue
-                                    else:
-                                        nl.insert(i+1, col_name)  # insert column following previous found column
-                                        collection.customViewFields = tuple(nl)
-                                        break
-
-                add_rm_col(removed, 'rm')
-                add_rm_col(added, 'add')
-
-            # We check if there are fields to empty
-            empty = field_constraints.get('empty', {})
-            empty_fields(event, empty)